more cleanup

+++ src/main/java/io/turntabl/model/metrics/G1GarbageCollectionDuration.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 
 public class G1GarbageCollectionDuration extends GarbageCollection{
 
-    public G1GarbageCollectionDuration(String name, String type, HashMap<String, Double> value, Long timestamp, Long interval, HashMap<String, String> attributes) {
+    public G1GarbageCollectionDuration(String name, String type, HashMap<String, Double> value, Long timestamp, int interval, HashMap<String, String> attributes) {
         super(name, type, value, timestamp, interval, attributes);
     }
