+++ src/main/java/io/turntabl/model/metrics/GCMinorDuration.java
@@ -3,7 +3,7 @@ package io.turntabl.model.metrics;
 import java.util.HashMap;
 
 public class GCMinorDuration extends GarbageCollection {
-    public GCMinorDuration(String name, String type, HashMap<String, Double> value, Long timestamp, Long interval, HashMap<String, String> attributes) {
+    public GCMinorDuration(String name, String type, HashMap<String, Double> value, Long timestamp, int interval, HashMap<String, String> attributes) {
         super(name, type, value, timestamp, interval, attributes);
     }
